feat: add export options toggle and TranscriptExporter component for session transcripts
